+++ packages/@eventual/project/src/project-type.ts
@@ -1,0 +1,4 @@
+export enum ProjectType {
+  AWS_CDK = "aws-cdk",
+  SST = "sst",
+}