+++ packages/create-eventual/src/sample-code.ts
@@ -66,36 +66,3 @@ const app = new App();
 new MyServiceStack(app, "${projectName}");
 `;
 }
-
-export function sampleCDKStack(projectName: string) {
-  return `import { Construct } from "constructs";
-import { CfnOutput, Stack, StackProps } from "aws-cdk-lib";
-import { Service } from "@eventual/aws-cdk";
-import path from "path";
-
-export interface MyServiceStackProps extends StackProps {}
-
-export class MyServiceStack extends Stack {
-  public readonly service: Service;
-
-  constructor(scope: Construct, id: string, props?: MyServiceStackProps) {
-    super(scope, id, props);
-
-    this.service = new Service(this, "${projectName}", {
-      name: "${projectName}",
-      entry: path.join(__dirname, "..", "..", "services", "src", "index.ts")
-    });
-
-    new CfnOutput(this, "my-service-api-endpoint", {
-      exportName: "my-service-api-endpoint",
-      value: this.service.api.gateway.url!,
-    });
-
-    new CfnOutput(this, "my-service-event-bus-arn", {
-      exportName: "my-service-api-endpoint",
-      value: this.service.events.bus.eventBusArn,
-    });
-  }
-}
-`;
-}